# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Chat With Your Docs
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type DocumentStatus = typeof DocumentStatus[keyof typeof DocumentStatus];


export const DocumentStatus = {
  UPLOADING: 'UPLOADING',
  UPLOADED: 'UPLOADED',
  PROCESSING: 'PROCESSING',
  EMBEDDING: 'EMBEDDING',
  READY: 'READY',
  FAILED: 'FAILED',
} as const;

export interface Document {
  id: string;
  filename: string;
  originalFilename: string;
  fileType: string;
  fileSize: number;
  /** @nullable */
  storagePath?: string | null;
  status: DocumentStatus;
  /** @nullable */
  pageCount?: number | null;
  /** @nullable */
  errorMessage?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface Conversation {
  id: string;
  title: string;
  /** @nullable */
  documentId: string | null;
  messageCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface DashboardSummary {
  totalDocuments: number;
  readyDocuments: number;
  processingDocuments: number;
  totalConversations: number;
  recentDocuments: Document[];
  recentConversations: Conversation[];
}

export type DocumentDetail = Document & {
  chunkCount: number;
};

export type DocumentInputFileType = typeof DocumentInputFileType[keyof typeof DocumentInputFileType];


export const DocumentInputFileType = {
  'application/pdf': 'application/pdf',
  'text/plain': 'text/plain',
} as const;

export interface DocumentInput {
  /** @minLength 1 */
  filename: string;
  /** @minLength 1 */
  originalFilename: string;
  fileType: DocumentInputFileType;
  /** @minimum 1 */
  fileSize: number;
  /** @minLength 1 */
  storagePath: string;
}

export interface DocumentUpdate {
  /**
     * @minLength 1
     * @maxLength 255
     */
  filename: string;
}

export type MessageRole = typeof MessageRole[keyof typeof MessageRole];


export const MessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface Source {
  chunkId: string;
  documentId: string;
  documentName: string;
  content: string;
  /** @nullable */
  pageNumber: number | null;
  similarity: number;
}

export interface Message {
  id: string;
  role: MessageRole;
  content: string;
  sources: Source[];
  createdAt: string;
}

export type ConversationDetail = Conversation & {
  messages: Message[];
};

export interface ConversationInput {
  /** @maxLength 120 */
  title?: string;
  /** @nullable */
  documentId?: string | null;
}

export interface ConversationUpdate {
  /**
     * @minLength 1
     * @maxLength 120
     */
  title: string;
}

export interface ChatInput {
  /**
     * @minLength 1
     * @maxLength 4000
     */
  question: string;
  /** @nullable */
  conversationId?: string | null;
  /** @nullable */
  documentId?: string | null;
}

export interface ChatResponse {
  answer: string;
  sources: Source[];
  conversationId: string;
  messageId: string;
}

export interface SearchInput {
  /**
     * @minLength 1
     * @maxLength 4000
     */
  query: string;
  /** @nullable */
  documentId?: string | null;
  /**
     * @minimum 1
     * @maximum 20
     */
  topK?: number;
}

export type SearchResultMetadata = { [key: string]: unknown };

export type SearchResult = Source & {
  metadata: SearchResultMetadata;
};

export type UploadRequestContentType = typeof UploadRequestContentType[keyof typeof UploadRequestContentType];


export const UploadRequestContentType = {
  'application/pdf': 'application/pdf',
  'text/plain': 'text/plain',
} as const;

export interface UploadRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  contentType: UploadRequestContentType;
}

export interface UploadResponse {
  uploadURL: string;
  objectPath: string;
}

/**
 * Not found
 */
export type NotFoundResponse = ErrorResponse;