# lib/api-zod/src/generated/types/chatResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Chat With Your Docs
 * OpenAPI spec version: 0.1.0
 */
import type { Source } from './source';

export interface ChatResponse {
  answer: string;
  sources: Source[];
  conversationId: string;
  messageId: string;
}