# lib/api-zod/src/generated/types/conversationDetail.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Chat With Your Docs
 * OpenAPI spec version: 0.1.0
 */
import type { Conversation } from './conversation';
import type { Message } from './message';

export type ConversationDetail = Conversation & {
  messages: Message[];
};