# lib/api-zod/src/generated/types/conversationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Chat With Your Docs
 * OpenAPI spec version: 0.1.0
 */

export interface ConversationInput {
  /** @maxLength 120 */
  title?: string;
  /** @nullable */
  documentId?: string | null;
}