# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Chat With Your Docs
 * OpenAPI spec version: 0.1.0
 */
import type { Conversation } from './conversation';
import type { Document } from './document';

export interface DashboardSummary {
  totalDocuments: number;
  readyDocuments: number;
  processingDocuments: number;
  totalConversations: number;
  recentDocuments: Document[];
  recentConversations: Conversation[];
}