# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Chat With Your Docs
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentStatus } from './documentStatus';

export interface Document {
  id: string;
  filename: string;
  originalFilename: string;
  fileType: string;
  fileSize: number;
  /** @nullable */
  storagePath?: string | null;
  status: DocumentStatus;
  /** @nullable */
  pageCount?: number | null;
  /** @nullable */
  errorMessage?: string | null;
  createdAt: Date;
  updatedAt: Date;
}