# lib/api-zod/src/generated/types/documentDetail.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Chat With Your Docs
 * OpenAPI spec version: 0.1.0
 */
import type { Document } from './document';

export type DocumentDetail = Document & {
  chunkCount: number;
};