# lib/api-zod/src/generated/types/searchInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Chat With Your Docs
 * OpenAPI spec version: 0.1.0
 */

export interface SearchInput {
  /**
     * @minLength 1
     * @maxLength 4000
     */
  query: string;
  /** @nullable */
  documentId?: string | null;
  /**
     * @minimum 1
     * @maximum 20
     */
  topK?: number;
}