# lib/api-zod/src/generated/types/uploadRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Chat With Your Docs
 * OpenAPI spec version: 0.1.0
 */
import type { UploadRequestContentType } from './uploadRequestContentType';

export interface UploadRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  contentType: UploadRequestContentType;
}